Separate unit tests for user waiver process
